agent: drop LastCheckedTime from status
The GVariant field this depended on was removed in the last iteration of
the rpm-ostree WIP. In any case, it wasn't exactly correct anyway. We
were conflating the check time, with the last time we did detect an
update. We could re-implement this later on using something similar to
what `rpm-ostree status` does, i.e. querying systemd itself.